requirements.txt
Flask==1.1.2
python-dotenv~=0.15.0
Jinja2==2.11.3
